refactor: organize Auth services and controller into feature-specific folders
