+++ GoTorz.Api/Services/Auth/AuthService.cs
@@ -1,7 +1,7 @@
 ﻿using GoTorz.Shared.Auth;
 using Microsoft.AspNetCore.Identity;
 
-namespace GoTorz.Api.Services
+namespace GoTorz.Api.Services.Auth
 {
     /// <summary>
     /// Implements authentication logic using ASP.NET Core Identity.
